Revert "add mathjax 3 as experimental toggle via query param (#1565)"
This reverts commit dec4bd7876953f516b15d48109670d12e2bb0abc.
