+++ public/index.html
@@ -38,6 +38,7 @@
     </script>
     <!-- End Google Optimize -->
 
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-MML-AM_HTMLorMML-full&delayStartupUntil=configured" async></script>
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
@@ -78,7 +79,7 @@
       s.src = '//js.pulseinsights.com/surveys.js';
       var f = d.getElementsByTagName('script')[0];
       f.parentNode.insertBefore(s, f);
-
+      
       pi('identify', 'PI-16384954');
       pi('get', 'surveys');
       })();
