+++ src/helpers/mathjax.spec.ts
@@ -1,48 +1,188 @@
-import { assertDocument, assertWindow } from '../app/utils';
 import { startMathJax, typesetMath } from './mathjax';
-import { startMathJax as startMathJax2, typesetMath as typesetMath2 } from './mathjax/v2';
-import { startMathJax as startMathJax3, typesetMath as typesetMath3 } from './mathjax/v3';
 
-jest.mock('./mathjax/v2', () => ({
-  startMathJax: jest.fn(),
-  typesetMath: jest.fn(),
-}));
+const debounce = () => new Promise((resolve) => setTimeout(resolve, 150));
 
-jest.mock('./mathjax/v3', () => ({
-  startMathJax: jest.fn(),
-  typesetMath: jest.fn(),
-}));
+const mockMathJax = () => ({
+  HTML: {
+    Cookie: {},
+  },
+  Hub: {
+    Config: jest.fn(),
+    Configured: jest.fn(),
+    Queue: (...fns: Array<() => void>) => fns.forEach((fn) => fn()),
+    Typeset: jest.fn(),
+  },
+});
 
-describe('v2', () => {
-  it('starts to the correct version', () => {
-    startMathJax();
-    expect(startMathJax2).toHaveBeenCalled();
+beforeEach(() => {
+  if (window) {
+    window.MathJax = mockMathJax();
+  }
+});
+afterEach(() => {
+  if (window) {
+    delete window.MathJax;
+  }
+});
+
+describe('typesetMath', () => {
+  it('does nothing for elements with no math', async() => {
+    if (!document || !window) {
+      expect(document).toBeTruthy();
+      expect(window).toBeTruthy();
+      return;
+    }
+    const element = document.createElement('div');
+
+    typesetMath(element);
+
+    await debounce();
+
+    expect(window.MathJax.Hub.Typeset).not.toHaveBeenCalledWith();
   });
 
-  it('calls the correct version', () => {
-    const element = assertDocument().createElement('div');
+  it('typesets the element if it contains mathml', async() => {
+    if (!document || !window) {
+      expect(document).toBeTruthy();
+      expect(window).toBeTruthy();
+      return;
+    }
+    const element = document.createElement('div');
+    element.appendChild(document.createElement('math'));
+
     typesetMath(element);
-    expect(typesetMath2).toHaveBeenCalled();
+
+    await debounce();
+
+    expect(window.MathJax.Hub.Typeset).toHaveBeenCalledWith(element);
+  });
+
+  it('debounces', async() => {
+    if (!document || !window) {
+      expect(document).toBeTruthy();
+      expect(window).toBeTruthy();
+      return;
+    }
+
+    const element = document.createElement('div');
+    element.appendChild(document.createElement('math'));
+
+    const element2 = document.createElement('div');
+    element2.appendChild(document.createElement('math'));
+
+    typesetMath(element);
+    typesetMath(element);
+    typesetMath(element);
+    typesetMath(element2);
+    typesetMath(element2);
+    typesetMath(element2);
+    typesetMath(element);
+    typesetMath(element2);
+    typesetMath(element);
+
+    await debounce();
+
+    expect(window.MathJax.Hub.Typeset).toHaveBeenCalledTimes(2);
+  });
+
+  it('typesets the element if it is data-math', async() => {
+    if (!document || !window) {
+      expect(document).toBeTruthy();
+      expect(window).toBeTruthy();
+      return;
+    }
+    const element = document.createElement('div');
+    const math1 = document.createElement('div');
+    math1.setAttribute('data-math', 'formula1');
+
+    const math2 = document.createElement('div');
+    math2.setAttribute('data-math', 'formula2');
+
+    element
+      .appendChild(math1)
+      .appendChild(math2)
+    ;
+
+    typesetMath(element);
+
+    await debounce();
+
+    expect(window.MathJax.Hub.Typeset).toHaveBeenCalledWith([math1, math2]);
+    expect(window.MathJax.Hub.Typeset).not.toHaveBeenCalledWith(element);
+  });
+
+  it('moves latex formulas inline', async() => {
+    if (!document || !window) {
+      expect(document).toBeTruthy();
+      expect(window).toBeTruthy();
+      return;
+    }
+    const element = document.createElement('div');
+    const math1 = document.createElement('div');
+    math1.setAttribute('data-math', 'formula1');
+
+    const math2 = document.createElement('span');
+    math2.setAttribute('data-math', 'formula2');
+
+    element
+      .appendChild(math1)
+      .appendChild(math2)
+    ;
+
+    typesetMath(element);
+
+    await debounce();
+
+    expect(math1.textContent).toEqual('\u200c\u200c\u200cformula1\u200c\u200c\u200c');
+    expect(math2.textContent).toEqual('\u200b\u200b\u200bformula2\u200b\u200b\u200b');
   });
 });
 
-describe('v3', () => {
-  beforeEach(() => {
-    Object.defineProperty(assertWindow(), 'location', {
-      value: {
-        search: 'mathjax-version=3',
-      },
-    });
-  });
+describe('startMathJax', () => {
+  it('loads config if mathjax is not defined', () => {
+    if (!window) {
+      expect(window).toBeTruthy();
+      return;
+    }
+    delete window.MathJax;
 
-  it('starts the correct version', () => {
     startMathJax();
-    expect(startMathJax3).toHaveBeenCalled();
+
+    expect(window.MathJax).toBeDefined();
   });
 
-  it('calls the correct version', () => {
-    const element = assertDocument().createElement('div');
-    typesetMath(element);
-    expect(typesetMath3).toHaveBeenCalled();
+  it('configs mathjax', () => {
+    if (!window) {
+      expect(window).toBeTruthy();
+      return;
+    }
+
+    startMathJax();
+
+    expect(window.MathJax.Hub.Config).toHaveBeenCalled();
+    expect(window.MathJax.Hub.Configured).toHaveBeenCalled();
+  });
+
+  describe('outside the browser', () => {
+    const windowBak = window;
+
+    beforeEach(() => {
+      delete (global as any).window;
+    });
+    afterEach(() => {
+      (global as any).window = windowBak;
+    });
+
+    it('throws', () => {
+      let message = null;
+
+      try {
+        startMathJax();
+      } catch (e) {
+        message = e.message;
+      }
+
+      expect(message).toBe('BUG: Window is undefined');
+    });
   });
 });
